build: fix circular dep between interface and l_node by merging (#20855)
PR Close #20855
