+++ modules/benchmarks/src/largetable/render3/table.ts
@@ -7,7 +7,7 @@
  */
 
 import {ɵC as C, ɵE as E, ɵT as T, ɵV as V, ɵb as b, ɵc as c, ɵcR as cR, ɵcr as cr, ɵdefineComponent as defineComponent, ɵdetectChanges as detectChanges, ɵe as e, ɵs as s, ɵt as t, ɵv as v} from '@angular/core';
-import {ComponentDef} from '@angular/core/src/render3/public_interfaces';
+import {ComponentDef} from '@angular/core/src/render3/definition_interfaces';
 
 import {TableCell, buildTable, emptyTable} from '../util';
