+++ packages/core/src/render3/definition_interfaces.ts
@@ -6,9 +6,9 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {RendererType2, Type} from '../core';
+import {RendererType2} from '../render/api';
+import {Type} from '../type';
 import {resolveRendererType2} from '../view/util';
-import {componentRefresh, diPublic} from './instructions';
 
 
 
@@ -78,7 +78,7 @@ export interface DirectiveDef<T> {
    * @param directiveIndex index of the directive in the containing template
    * @param elementIndex index of an host element for a given directive.
    */
-  r(this: DirectiveDef<T>, directiveIndex: number, elementIndex: number): void;
+  r(directiveIndex: number, elementIndex: number): void;
 }
 
 export interface ComponentDef<T> extends DirectiveDef<T> {
@@ -93,7 +93,7 @@ export interface ComponentDef<T> extends DirectiveDef<T> {
    * @param directiveIndex index of the directive in the containing template
    * @param elementIndex index of an host element for a given component.
    */
-  r(this: ComponentDef<T>, directiveIndex: number, elementIndex: number): void;
+  r(directiveIndex: number, elementIndex: number): void;
 
   /**
    * The tag name which should be used by the component.
@@ -120,7 +120,7 @@ export interface ComponentDef<T> extends DirectiveDef<T> {
 export interface DirectiveDefArgs<T> {
   type: Type<T>;
   factory: () => T;
-  refresh?: (this: DirectiveDef<T>, directiveIndex: number, elementIndex: number) => void;
+  refresh?: (directiveIndex: number, elementIndex: number) => void;
   inputs?: {[P in keyof T]?: string};
   outputs?: {[P in keyof T]?: string};
   methods?: {[P in keyof T]?: string};
@@ -130,80 +130,10 @@ export interface DirectiveDefArgs<T> {
 export interface ComponentDefArgs<T> extends DirectiveDefArgs<T> {
   tag: string;
   template: ComponentTemplate<T>;
-  refresh?: (this: ComponentDef<T>, directiveIndex: number, elementIndex: number) => void;
+  refresh?: (directiveIndex: number, elementIndex: number) => void;
   features?: ComponentDefFeature[];
   rendererType?: RendererType2;
 }
 
 export type DirectiveDefFeature = <T>(directiveDef: DirectiveDef<T>) => void;
 export type ComponentDefFeature = <T>(directiveDef: DirectiveDef<T>) => void;
-
-/**
- * Create a component definition object.
- *
- *
- * # Example
- * ```
- * class MyDirective {
- *   // Generated by Angular Template Compiler
- *   // [Symbol] syntax will not be supported by TypeScript until v2.7
- *   static [COMPONENT_DEF_SYMBOL] = defineComponent({
- *     ...
- *   });
- * }
- * ```
- */
-export function defineComponent<T>(componentDefinition: ComponentDefArgs<T>): ComponentDef<T> {
-  const def = <ComponentDef<any>>{
-    type: componentDefinition.type,
-    diPublic: null,
-    n: componentDefinition.factory,
-    tag: (componentDefinition as ComponentDefArgs<T>).tag || null !,
-    template: (componentDefinition as ComponentDefArgs<T>).template || null !,
-    r: componentDefinition.refresh || componentRefresh,
-    inputs: invertObject(componentDefinition.inputs),
-    outputs: invertObject(componentDefinition.outputs),
-    methods: invertObject(componentDefinition.methods),
-    rendererType: resolveRendererType2(componentDefinition.rendererType) || null,
-  };
-  const feature = componentDefinition.features;
-  feature && feature.forEach((fn) => fn(def));
-  return def;
-}
-
-export function NgOnChangesFeature<T>(definition: DirectiveDef<T>) {
-  // TODO: implement. See: https://app.asana.com/0/443577627818617/465170715764659
-}
-
-export function PublicFeature<T>(definition: DirectiveDef<T>) {
-  definition.diPublic = diPublic;
-}
-
-const EMPTY = {};
-
-/** Swaps the keys and values of an object. */
-function invertObject(obj: any): any {
-  if (obj == null) return EMPTY;
-  const newObj: any = {};
-  for (let minifiedKey in obj) {
-    newObj[obj[minifiedKey]] = minifiedKey;
-  }
-  return newObj;
-}
-
-/**
- * Create a directive definition object.
- *
- * # Example
- * ```
- * class MyDirective {
- *   // Generated by Angular Template Compiler
- *   // [Symbol] syntax will not be supported by TypeScript until v2.7
- *   static [DIRECTIVE_DEF_SYMBOL] = defineDirective({
- *     ...
- *   });
- * }
- * ```
- */
-export const defineDirective = defineComponent as<T>(directiveDefinition: DirectiveDefArgs<T>) =>
-    DirectiveDef<T>;
